requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.0.0
daphne==4.0.0
channels==4.0.0
channels-redis==4.1.0
celery==5.3.0
redis==5.0.0
psycopg2-binary==2.9.0
gunicorn==21.2.0
djangorestframework-simplejwt
django_celery_beat
django-redis